+++ src/main/java/svenhjol/charmony/core/base/Feature.java
@@ -9,6 +9,7 @@ import java.util.*;
  * Holder for sidedFeature instances.
  * Use get(side) to get a specific sided feature and sides() for all sided features.
  */
+@SuppressWarnings({"DeprecatedIsStillUsed", "unused"})
 public final class Feature {
     private final Log log;
     private final Mod mod;
@@ -35,6 +36,11 @@ public final class Feature {
         return features.values().stream().anyMatch(SidedFeature::canBeDisabled);
     }
 
+    public boolean canBeDisabledInConfig() {
+        return features.values().stream().anyMatch(SidedFeature::canBeDisabledInConfig);
+    }
+
+    @Deprecated(since = "1.20.0")
     public boolean showInConfig() {
         return features.values().stream().anyMatch(SidedFeature::showInConfig);
     }