+++ src/main/java/svenhjol/charmony/core/client/features/core/Core.java
@@ -18,6 +18,6 @@ public final class Core extends SidedFeature {
     }
 
     public static Core feature() {
-        return Charmony.instance().sidedFeature(Core.class);
+        return Mod.getSidedFeature(Core.class);
     }
 }