+++ src/main/java/svenhjol/charmony/core/common/features/advancements/Registers.java
@@ -14,8 +14,6 @@ public class Registers extends Setup<Advancements> {
 
     @Override
     public Runnable boot() {
-        return () -> {
-            ServerPlayConnectionEvents.JOIN.register(feature().handlers::playerJoin);
-        };
+        return () -> ServerPlayConnectionEvents.JOIN.register(feature().handlers::playerJoin);
     }
 }