+++ src/main/java/svenhjol/charmony/core/common/features/core/Registers.java
@@ -15,8 +15,6 @@ public class Registers extends Setup<Core> {
 
     @Override
     public Runnable boot() {
-        return () -> {
-            ServerPlayConnectionEvents.JOIN.register(feature().handlers::playerJoin);
-        };
+        return () -> ServerPlayConnectionEvents.JOIN.register(feature().handlers::playerJoin);
     }
 }